+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## dbt-databricks 1.11.7 (TBD)
 
+### Features
+
+- Enable Notebook scoped python packages installation
+- 
 ### Fixes
 
 - Fix column order mismatch in microbatch and replace_where incremental strategies by using INSERT BY NAME syntax ([#1338](https://github.com/databricks/dbt-databricks/issues/1338))